Move pipe tests to set.spec
